refactor: organize library
