+++ src/util.rs
@@ -20,81 +20,6 @@ use std::sync::Arc;
 // use sysinfo::{ get_current_pid, CpuExt, ProcessExt, System, SystemExt, UserExt };
 use std::path::Path;
 
-pub trait OSInfo: Send + Sync {
-    fn sys_font(&self) -> Option<String> {
-        None
-    }
-
-    fn cursor(&self) -> Option<String> {
-        None
-    }
-    fn terminal(&self) -> Option<String> {
-        None
-    }
-    fn term_font(&self) -> Option<String> {
-        None
-    }
-    fn gpus(&self) -> Vec<Arc<str>> {
-        Vec::new()
-    }
-    fn memory(&self) -> Option<String> {
-        None
-    }
-    fn disks(&self) -> Vec<(String, String)> {
-        Vec::new()
-    }
-    fn battery(&self) -> Option<String> {
-        None
-    }
-    fn locale(&self) -> Option<String> {
-        None
-    }
-    fn theme(&self) -> Option<String> {
-        None
-    }
-    fn icons(&self) -> Option<String> {
-        None
-    }
-    fn os(&self) -> Option<String> {
-        None
-    }
-    fn id(&self) -> Arc<str>;
-    fn uptime(&self) -> Option<String>;
-    fn ip(&self) -> Vec<Arc<str>>;
-    fn displays(&self) -> Vec<Arc<str>> {
-        Vec::new()
-    }
-
-    fn hostname(&self) -> Option<Arc<str>>;
-
-    fn machine(&self) -> Option<String> {
-        None
-    }
-
-    fn kernel(&self) -> Option<String> {
-        None
-    }
-
-    fn wm(&self) -> Option<String> {
-        None
-    }
-
-    fn de(&self) -> Option<String> {
-        None
-    }
-
-    fn shell(&self) -> Option<String> {
-        None
-    }
-    fn cpu(&self) -> Option<String> {
-        None
-    }
-
-    fn username(&self) -> Option<Arc<str>> {
-        None
-    }
-}
-
 /// .
 ///
 /// # Errors
